reworked spotifyless animations
