+++ animation_data/Theater.json
@@ -1,18 +1,14 @@
 {
   "name": "Theater",
-  "model": {
-    "animation": "Mirror",
-    "config": {
-      "sub": {
-        "animation": "Theater",
-        "config": {
-          "delay": 0.1
-        }
-      },
-      "inverse": [
-        true,
-        false
-      ]
+  "animation": {
+    "name": "Mirror",
+    "inverse": [
+      true,
+      false
+    ],
+    "animation": {
+      "name": "Theater",
+      "delay": 0.1
     }
   }
 }
