+++ api/animation/decorators.py
@@ -3,16 +3,15 @@ from typing import Callable
 
 import numpy as np
 
-from ...color import Color
-from .abstract import Animation
+from ..color import Color
+from .abstract import AbstractAnimation
 
 
 def on_change(function: Callable[..., np.ndarray]) -> Callable:
     @wraps(function)
-    def wrapper(self: Animation, progress: float, xy: np.ndarray, **kwargs):
-        if not np.array_equal(xy, getattr(self, "_xy", None)) or getattr(self, "_change_trigger", False):
+    def wrapper(self: AbstractAnimation, progress: float, xy: np.ndarray, **kwargs):
+        if not np.array_equal(xy, getattr(self, "_xy", None)):
             setattr(self, "_xy", xy)
-            setattr(self, "_change_trigger", False)
             self.change_callback(xy)
         return function(self, progress, xy, **kwargs)
 
@@ -21,7 +20,7 @@ def on_change(function: Callable[..., np.ndarray]) -> Callable:
 
 def save_previous(function: Callable[..., np.ndarray]) -> Callable:
     @wraps(function)
-    def wrapper(self: Animation, progress: float, xy: np.ndarray, **kwargs):
+    def wrapper(self: AbstractAnimation, progress: float, xy: np.ndarray, **kwargs):
         previous = getattr(self, "_prev", None)
         if previous is None or len(previous) != len(xy):
             previous = np.full(len(xy), Color())
