--- api/animation/spotifyless/__init__.py
@@ -1,3 +1,0 @@
-from .pride import Pride
-from .rainbow import Rainbow
-from .theater import Theater
